requirements.txt
# APEX-ULTRA™ v15.0 AGI COSMOS - Dependencies
# Core system dependencies for production-ready AGI system

# =============================================================================
# CORE PYTHON DEPENDENCIES
# =============================================================================

# Async programming and concurrency
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=0.8.0
asyncio-mqtt>=0.11.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine learning and AI
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.15.0
openai>=0.27.0
anthropic>=0.3.0
langchain>=0.0.200
sentence-transformers>=2.2.0

# Database and storage
sqlalchemy>=1.4.0
alembic>=1.7.0
redis>=4.0.0
pymongo>=4.0.0
elasticsearch>=8.0.0
influxdb>=5.3.0

# Web and API frameworks
fastapi>=0.78.0
uvicorn>=0.17.0
requests>=2.28.0
httpx>=0.23.0
websockets>=10.0
flask>=2.2.0
django>=4.0.0

# =============================================================================
# API INTEGRATION DEPENDENCIES
# =============================================================================

# Social media APIs
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
tweepy>=4.10.0
praw>=7.5.0
python-instagram>=1.5.0
linkedin-api>=2.0.0
facebook-sdk>=3.1.0

# Payment and e-commerce
stripe>=4.0.0
paypalrestsdk>=1.13.0
shopify>=8.0.0
braintree>=4.0.0

# Cloud services
boto3>=1.24.0
azure-storage-blob>=12.0.0
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.15.0

# Communication services
twilio>=7.15.0
sendgrid>=6.9.0
mailchimp-marketing>=3.0.0
slack-sdk>=3.15.0
discord.py>=2.0.0

# =============================================================================
# DATA PROCESSING AND ANALYTICS
# =============================================================================

# Data manipulation
polars>=0.15.0
vaex>=4.13.0
dask>=2022.1.0
xarray>=2022.1.0

# Time series analysis
statsmodels>=0.13.0
prophet>=1.1.0
arch>=5.3.0
tslearn>=0.5.0

# Natural language processing
nltk>=3.7
spacy>=3.4.0
textblob>=0.17.0
gensim>=4.2.0
wordcloud>=1.8.0

# Computer vision
opencv-python>=4.6.0
pillow>=9.0.0
imageio>=2.15.0
albumentations>=1.2.0

# =============================================================================
# MONITORING AND OBSERVABILITY
# =============================================================================

# System monitoring
psutil>=5.8.0
prometheus-client>=0.13.0
grafana-api>=1.0.0
datadog>=0.44.0

# Logging and tracing
structlog>=21.5.0
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
jaeger-client>=4.8.0

# Performance profiling
memory-profiler>=0.60.0
line-profiler>=3.5.0
py-spy>=0.3.0

# =============================================================================
# SECURITY AND ENCRYPTION
# =============================================================================

# Cryptography and security
cryptography>=37.0.0
bcrypt>=4.0.0
passlib>=1.7.4
python-jose>=3.3.0
pyjwt>=2.4.0

# Security scanning
bandit>=1.7.0
safety>=1.10.0
semgrep>=0.100.0

# =============================================================================
# TESTING AND QUALITY ASSURANCE
# =============================================================================

# Testing frameworks
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
hypothesis>=6.50.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950
pylint>=2.12.0

# Type checking
types-requests>=2.28.0
types-PyYAML>=6.0.0
types-psutil>=5.8.0

# =============================================================================
# DEVELOPMENT AND DEPLOYMENT
# =============================================================================

# Development tools
pre-commit>=2.19.0
tox>=3.25.0
coverage>=6.3.0

# Containerization
docker>=6.0.0
kubernetes>=26.0.0

# Configuration management
python-dotenv>=0.19.0
pydantic>=1.9.0
pydantic-settings>=1.0.0
click>=8.1.0
typer>=0.6.0

# =============================================================================
# UTILITIES AND HELPERS
# =============================================================================

# Date and time
python-dateutil>=2.8.0
pytz>=2022.1
arrow>=1.2.0

# File handling
pathlib2>=2.3.0
watchdog>=2.1.0
python-magic>=0.4.24

# Network and HTTP
urllib3>=1.26.0
certifi>=2022.5.0
charset-normalizer>=2.0.0

# Serialization
pyyaml>=6.0
toml>=0.10.0
msgpack>=1.0.0
orjson>=3.6.0

# =============================================================================
# OPTIONAL DEPENDENCIES (UNCOMMENT AS NEEDED)
# =============================================================================

# GPU acceleration (uncomment if using CUDA)
# cupy-cuda11x>=11.0.0
# torch-cuda>=1.10.0

# Advanced visualization (uncomment for enhanced dashboards)
# dash>=2.4.0
# streamlit>=1.15.0
# bokeh>=2.4.0
# holoviews>=1.15.0

# Blockchain integration (uncomment for crypto features)
# web3>=5.30.0
# eth-account>=0.8.0
# bitcoin>=1.1.0

# IoT and hardware (uncomment for IoT features)
# paho-mqtt>=1.6.0
# pyserial>=3.5
# gpiozero>=1.6.0

# =============================================================================
# VERSION CONSTRAINTS
# =============================================================================

# Ensure compatibility
setuptools>=65.0.0
wheel>=0.37.0
pip>=22.0.0

# =============================================================================
# DEVELOPMENT DEPENDENCIES (FOR CONTRIBUTORS)
# =============================================================================

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0

# Code formatting and linting
autopep8>=1.7.0
yapf>=0.32.0
prospector>=1.7.0

# =============================================================================
# PRODUCTION DEPENDENCIES
# =============================================================================

# Production server
gunicorn>=20.1.0
uvloop>=0.16.0

# Caching and performance
aioredis>=2.0.0
memcached>=1.59

# Background tasks
celery>=5.2.0
redis>=4.0.0

# =============================================================================
# MONITORING AND ALERTING
# =============================================================================

# Application monitoring
sentry-sdk>=1.9.0
newrelic>=8.0.0

# Health checks
healthcheck>=1.3.0

# =============================================================================
# BACKUP AND RECOVERY
# =============================================================================

# Backup utilities
boto3>=1.24.0
azure-storage-blob>=12.0.0
google-cloud-storage>=2.0.0

# Compression
lz4>=3.1.0
zstandard>=0.18.0

# =============================================================================
# COMPATIBILITY NOTES
# =============================================================================

# Python version compatibility
# This system requires Python 3.8+ for full functionality
# Some dependencies may have specific version requirements
# Always test in a virtual environment before deployment

# Platform compatibility
# Most dependencies work on Linux, macOS, and Windows
# GPU acceleration requires CUDA-compatible hardware
# IoT features require appropriate hardware interfaces 